Rename 'onstack' to 'optional' and implement 'onstack' on top of it
Rename the 'onstack' policy to 'optional', because it's closest in
behaviour to std::optional due to it supporting a null-state.

Then implement a new 'onstack' on top of that, which does not permit a
null-state. The only difference it needs to accomplish this is deleting
the default constructor of 'optional' (which is the only place that a
null-state gets created).