Add 'always_storage': non-optional storage policy for 'inplace'
The `always_storage` storage policy, unlike `storage` does not support a
null-state. Additionally it of not having to support this by eliminating
the `traits` pointer from the object. As a result its size and alignment
can be exactly equal to that of the implementation. Meaning there's no
more, per object, memory overhead.