inplace: optimize: remove runtime branch when no null state is possible
Check whether the null-state is impossible before checking whether we
are not currently in the null state. Because the former is known at
compile-time, this allows optimizing the second check away if it is
indeed impossible.

This eliminates a lot of code (compiler-generated assembly) when the
inplace policy hasn't got a null-state.