inplace: track existence directly
I.e. don't use a traits pointer anymore that, when not null, signals the
existence of an implementation. Instead we store the existence as a
boolean flag. This can either be a real 'bool' or a type convertible to
'bool'. In this case we express existing always using an 'exists_always'
type that's convertible to 'bool', is always true, and throws when
attempting to assign false.

Furthermore we now use Boost's compressed_pair for the empty-base
optimization instead of doing it ourselves: this makes the code a lot
cleaner.