Add an RAII deallocation guard
Instead of wrapping code in try-catch blocks to ensure deallocation of
memory even in the face of exceptions, lets use RAII for that. I.e. all
memory is now deallocated through 'dealloc_guard'. It gets ownership of
the memory (but not the object in that) before calling the construct or
destructor to ensure that it's deallocated even if those throw. After
succesful construction however, we transfer ownership elsewhere.