Make traits::emplace more closely resemble policy::emplace
Make it get the derived_type by a template parameter instead of deriving
it from the allocator. Require the allocator to belong to the
derived_type instead.

Also take the allocator by non-const l-value reference, because we don't
want to take ownership, but we do need it to be non-const.